requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
